+++ Editor/Scripts/Drawers/DropdownAttributeDrawers/PropertyDropdownDrawer.cs
@@ -3,7 +3,6 @@ using UnityEngine;
 using UnityEditor;
 using UnityEngine.UIElements;
 using UnityEditor.UIElements;
-using System.Collections.Generic;
 using EditorAttributes.Editor.Utility;
 
 namespace EditorAttributes.Editor
@@ -17,7 +16,7 @@ namespace EditorAttributes.Editor
 
 			var root = new VisualElement();
 			var propertyField = CreatePropertyField(property);
-			var errorBox = new HelpBox("The PropertyDropdown Attribute can only be attached to objects deriving from <b>Component</b> or <b>ScriptableObject</b>", HelpBoxMessageType.Error);
+			var errorBox = new HelpBox("The PropertyDropdown Attribute can only be attached on to <b>UnityEngine.Object</b> types", HelpBoxMessageType.Error);
 
 			ApplyBoxStyle(root);
 
@@ -59,22 +58,7 @@ namespace EditorAttributes.Editor
 				return;
 			}
 
-			if (IsFieldOfType(fieldType, typeof(Component)))
-			{
-				var component = property.objectReferenceValue as Component;
-
-				root.Add(CreatePropertyFoldout(new SerializedObject(component), property));
-			}
-			else if (IsFieldOfType(fieldType, typeof(ScriptableObject)))
-			{
-				var scriptableObject = property.objectReferenceValue as ScriptableObject;
-
-				root.Add(CreatePropertyFoldout(new SerializedObject(scriptableObject), property));
-			}
-			else
-			{
-				root.Add(errorBox);
-			}
+			root.Add(CreatePropertyFoldout(new SerializedObject(property.objectReferenceValue), property));
 		}
 
 		private Foldout CreatePropertyFoldout(SerializedObject serializedObject, SerializedProperty serializedProperty)
@@ -89,59 +73,29 @@ namespace EditorAttributes.Editor
 
 			ApplyBoxStyle(foldout);
 
-			foldout.style.unityFontStyleAndWeight = FontStyle.Bold;
 			foldout.style.paddingLeft = 15f;
 
-			using (var property = serializedObject.GetIterator())
-			{
-				if (property.NextVisible(true))
-				{
-					do
-					{
-						if (property.name.Equals("m_Script", StringComparison.Ordinal)) // Exclude the field containing the script reference
-							continue;
-
-						var field = ReflectionUtility.FindField(property.name, property.serializedObject.targetObject);
-
-						var propertyField = CreatePropertyField(property);
-
-						propertyField.style.unityFontStyleAndWeight = FontStyle.Normal;
-
-						foldout.Add(propertyField);
-					}
-					while (property.NextVisible(false));
-				}
-			}
+			foldout.Add(new InspectorElement(serializedObject));
 
 			serializedObject.ApplyModifiedProperties();
 
 			foldout.RegisterValueChangedCallback((callback) => EditorPrefs.SetBool(foldoutSaveKey, callback.newValue));
+
+			ExecuteLater(foldout, () =>
+			{
+				foldout.Q<Label>(className: Foldout.textUssClassName).style.unityFontStyleAndWeight = FontStyle.Bold;
+				foldout.Q<ObjectField>("unity-input-m_Script")?.parent.RemoveFromHierarchy();
+			});
 
 			return foldout;
 		}
 
 		private Type GetValidPropertyType(SerializedProperty property)
 		{
-			var validProperty = ReflectionUtility.IsPropertyCollection(property) ? GetCollectionProperty(property) : property;
+			var validProperty = IsPropertyCollection(property) ? GetCollectionProperty(property) : property;
 			var memberInfo = ReflectionUtility.GetValidMemberInfo(validProperty.name, validProperty);
 
 			return ReflectionUtility.GetMemberInfoType(memberInfo);
-		}
-
-		private bool IsFieldOfType(Type fieldType, Type typeToCheck)
-		{
-			if (fieldType.IsArray)
-			{
-				return typeToCheck.IsAssignableFrom(fieldType.GetElementType());
-			}
-			else if (fieldType.IsGenericType && fieldType.GetGenericTypeDefinition() == typeof(List<>))
-			{
-				return typeToCheck.IsAssignableFrom(fieldType.GetGenericArguments()[0]);
-			}
-			else
-			{
-				return typeToCheck.IsAssignableFrom(fieldType);
-			}
 		}
 	}
 }